[JSC] Expensive header files slow full build https://bugs.webkit.org/show_bug.cgi?id=312607 rdar://175039075
Reviewed by Yusuke Suzuki.

Move ArrayBufferContents constructor bodies from ArrayBuffer.h to
ArrayBuffer.cpp and remove expensive includes of Options.h from
BufferMemoryHandle.h and MacroAssembler.h from WasmContext.h.
These headers are transitively included across many translation
units, so uninlining the constructors and removing the includes
reduces full build time.

This change improves a full build of `WebKit + Tools` by ~2m 30s on Mac Studio
M3 Ultra from ~15m to ~12m 30s.

* Source/JavaScriptCore/runtime/ArrayBuffer.cpp:
(JSC::ArrayBufferContents::ArrayBufferContents):
* Source/JavaScriptCore/runtime/ArrayBuffer.h:
* Source/JavaScriptCore/runtime/BufferMemoryHandle.cpp:
(JSC::BufferMemoryManager::BufferMemoryManager):
* Source/JavaScriptCore/runtime/BufferMemoryHandle.h:
* Source/JavaScriptCore/wasm/WasmContext.h:

Canonical link: https://commits.webkit.org/311631@main
