+++ Source/JavaScriptCore/wasm/WasmContext.h
@@ -29,7 +29,6 @@
 
 #if ENABLE(WEBASSEMBLY)
 
-#include <JavaScriptCore/MacroAssembler.h>
 #include <cstdint>
 #include <wtf/Lock.h>
 #include <wtf/UniqueArray.h>
